+++ R/install_translations.R
@@ -76,18 +76,32 @@ install_translations <- function(pkg = ".", verbose = getOption("verbose")) {
         dest <- file.path(inst_po_dir, lang, "LC_MESSAGES")
         dir.create(dest, FALSE, TRUE)
         dest <- file.path(dest, sprintf("R-%s.mo", pkg$package))
-        if (system(paste("msgfmt -c --statistics -o", shQuote(dest), shQuote(f))) != 0L) 
+        if (isTRUE(verbose)) {
+            cmd <- paste("msgfmt -v -o", shQuote(dest), shQuote(f))
+        } else {
+            cmd <- paste("msgfmt -v -o", shQuote(dest), shQuote(f))
+        }
+        if (system(cmd) != 0L) {
             warning(sprintf("running msgfmt on %s failed", basename(f)), immediate. = TRUE)
+        }
     }
+    # handle UTF-8 locale
     if (l10n_info()[["UTF-8"]]) {
-        lang <- "en@quot"
-        f <- tempfile()
-        dest <- file.path(inst_po_dir, lang, "LC_MESSAGES")
-        dir.create(dest, FALSE, TRUE)
-        dest <- file.path(dest, sprintf("R-%s.mo", pkg$package))
-        cmd <- paste("msgfmt -c --statistics -o", shQuote(dest), shQuote(f))
-        if (system(cmd) != 0L) 
-            warning(sprintf("running msgfmt on %s failed", basename(f)), immediate. = TRUE)
+        f1 <- make_enquote_translation(pkg = pkg, domain = "R", verbose = verbose)
+        f2 <- tempfile()
+        cmd <- paste("msgconv -t UTF-8 -o", f2, f1)
+        if (system(cmd) != 0L) {
+            stop("running msgconv on 'R-en@quot' UTF-8 localization failed", domain = NA)
+        }
+        dest <- file.path(inst_po_dir, "en@quot", "LC_MESSAGES", sprintf("R-%s.mo", pkg$package))
+        if (isTRUE(verbose)) {
+            cmd <- paste("msgfmt -v -o", shQuote(dest), shQuote(f2))
+        } else {
+            cmd <- paste("msgfmt -o", shQuote(dest), shQuote(f2))
+        }
+        if (system(cmd) != 0L) {
+            warning(sprintf("running msgfmt on 'R-en@quot' UTF-8 localization failed"), immediate. = TRUE)
+        }
     }
     
     # C-level messages
@@ -103,19 +117,33 @@ install_translations <- function(pkg = ".", verbose = getOption("verbose")) {
         dest <- file.path(inst_po_dir, lang, "LC_MESSAGES")
         dir.create(dest, FALSE, TRUE)
         dest <- file.path(dest, paste0("R-", pkg$package, ".mo"))
-        if (system(paste("msgfmt -c --statistics -o", shQuote(dest), shQuote(f))) != 0L) 
+        if (isTRUE(verbose)) {
+            cmd <- paste("msgfmt -v -o", shQuote(dest), shQuote(f))
+        } else {
+            cmd <- paste("msgfmt -o", shQuote(dest), shQuote(f))
+        }
+        if (system(cmd) != 0L) {
             warning(sprintf("running msgfmt on %s failed", basename(f)))
+        }
     }
-    if (l10n_info()[["UTF-8"]]) {
-        lang <- "en@quot"
-        f <- tempfile()
-        dest <- file.path(inst_po_dir, lang, "LC_MESSAGES")
-        dir.create(dest, FALSE, TRUE)
-        dest <- file.path(dest, sprintf("%s.mo", pkg$package))
-        cmd <- paste("msgfmt -c --statistics -o", shQuote(dest), shQuote(f))
-        if (system(cmd) != 0L) 
-            warning(sprintf("running msgfmt on %s failed", basename(f)))
+    if (length(po_files) && l10n_info()[["UTF-8"]]) {
+        f1 <- make_enquote_translation(pkg = pkg, domain = "C", verbose = verbose)
+        f2 <- tempfile()
+        cmd <- paste("msgconv -t UTF-8 -o", f2, f1)
+        if (system(cmd) != 0L) {
+            stop("running msgconv on 'en@quot' UTF-8 localization failed", domain = NA)
+        }
+        dest <- file.path(inst_po_dir, "en@quot", "LC_MESSAGES", sprintf("%s.mo", pkg$package))
+        if (isTRUE(verbose)) {
+            cmd <- paste("msgfmt -v -o", shQuote(dest), shQuote(f2))
+        } else {
+            cmd <- paste("msgfmt -o", shQuote(dest), shQuote(f2))
+        }
+        if (system(cmd) != 0L) {
+            warning(sprintf("running msgfmt on 'en@quot' UTF-8 localization failed"), immediate. = TRUE)
+        }
     }
+    
     return(invisible(TRUE))
 }